+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Create distinct file types to detect pipeline and task files
+- Add new file actions and templates
 
 ### Changed